# First working commit
The application loads data, produces a table of words sorted by frequency and is capable of reading the value of the word on a clicked row.

It also filters by directorate.

Now I will allow the clicked word to bring back all of the comments in that area which contain that word.